+++ server/db.go
@@ -8,7 +8,6 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
-	"sync"
 	"time"
 
 	_ "modernc.org/sqlite"
@@ -68,8 +67,8 @@ func backupWithVacuumInto(ctx context.Context, db DB, dir string) error {
 		return err
 	}
 
-	// If a backup for this date already exists, remove it so VACUUM INTO can recreate it.
-	// Possibly not great - May want to filter by quantity or age instead.
+	// If a backup for today exists, remove it and recreate it.
+	// Todo: filter by quantity or age instead.
 	if err := os.Remove(filename); err != nil && !os.IsNotExist(err) {
 		return err
 	}
@@ -117,70 +116,4 @@ func insertSnapshot(db DB, clicksA, clicksB, views int64) error {
 		VALUES (?,?,?,?)`,
 		time.Now().UTC().Unix(), clicksA, clicksB, views)
 	return err
-}
-
-// ---------- Point Broadcasts -------------
-type Broadcaster struct {
-	mu        sync.Mutex
-	listeners map[chan Point]struct{}
-}
-
-func NewBroadcaster() *Broadcaster {
-	return &Broadcaster{listeners: make(map[chan Point]struct{})}
-}
-
-func (b *Broadcaster) Subscribe() chan Point {
-	ch := make(chan Point, 100) // Keep buffer ?
-	b.mu.Lock()
-	b.listeners[ch] = struct{}{}
-	b.mu.Unlock()
-	return ch
-}
-
-func (b *Broadcaster) Unsubscribe(ch chan Point) {
-	b.mu.Lock()
-	delete(b.listeners, ch)
-	b.mu.Unlock()
-	close(ch)
-}
-
-func (b *Broadcaster) Publish(p Point) {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-	for ch := range b.listeners {
-		select { // Default to prevent lock ?
-		case ch <- p:
-			// noop
-		default:
-			// noop
-		}
-	}
-}
-
-func (app *App) sendPeriodicBroadcasts() {
-	if !app.configuration.broadcastEnabled() {
-		return
-	}
-	go func() {
-		ticker := time.NewTicker(app.configuration.broadcastInterval) // Source from config
-		defer ticker.Stop()
-
-		var previousClickACount, previousClickBCount, previousViewCount int64
-		for range ticker.C {
-			currentClicksA := app.clicksA.Load()
-			currentClicksB := app.clicksB.Load()
-			currentViews := app.views.Load()
-			if currentClicksA == previousClickACount &&
-				currentClicksB == previousClickBCount &&
-				currentViews == previousViewCount {
-				continue
-			}
-			app.broadcaster.Publish(Point{
-				Ts:      time.Now().UTC().Unix(),
-				ClicksA: currentClicksA,
-				ClicksB: currentClicksB,
-			})
-			previousClickACount, previousClickBCount, previousViewCount = currentClicksA, currentClicksB, currentViews
-		}
-	}()
 }
