+++ server/templates/home.tmpl.html
@@ -26,19 +26,7 @@
     </div>
 
     <div id="modal-container" data-show="$showModal">
-      <div id="modal-content">
-        <!-- <h2>Metrics</h2>
-        <div class="range-buttons" style="margin-bottom:1rem; display:flex; gap:0.5rem; flex-wrap:wrap;">
-          <button data-range="5m">5m</button>
-          <button data-range="1h">1h</button>
-          <button data-range="1d">1d</button>
-          <button data-range="2d">2d</button>
-          <button data-range="1w">1w</button>
-          <button data-range="all">All‑Time</button>
-        </div>
-        <canvas id="mChart"></canvas>
-        <a href="#" data-on-click="@get('metrics/toggle')">Hide</a> -->
-      </div>
+      <div id="modal-content"></div>
     </div>
   </body>
 </html>
